requirements.txt
pandas
streamlit
mysql-connector-python
sqlalchemy